Refactor and organise classes into packages
