+++ src/main/java/waz/command/Command.java
@@ -1,4 +1,9 @@
-import java.util.ArrayList;
+package waz.command;
+
+import waz.task.TaskList;
+import waz.exception.WazException;
+import waz.storage.Storage;
+import waz.ui.Ui;
 
 public abstract class Command {
     protected String argument;
