+++ src/main/java/waz/exception/WazException.java
@@ -1,3 +1,5 @@
+package waz.exception;
+
 public class WazException extends Exception{
     public WazException(String message) {
         super(message);
