+++ src/main/java/waz/task/Deadline.java
@@ -1,7 +1,8 @@
+package waz.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
-import java.util.Locale;
 
 public class Deadline extends Task {
 
@@ -17,8 +18,8 @@ public class Deadline extends Task {
     }
 
     /**
-     * Format the Task object into String to be saved in the file
-     * @return a formatted Task string
+     * Format the waz.task.Task object into String to be saved in the file
+     * @return a formatted waz.task.Task string
      */
     @Override
     public String toDataString() {
