+++ src/main/java/waz/task/Event.java
@@ -1,3 +1,5 @@
+package waz.task;
+
 public class Event extends Task {
 
     protected String from;
@@ -10,8 +12,8 @@ public class Event extends Task {
     }
 
     /**
-     * Format the Task object into String to be saved in the file
-     * @return a formatted Task string
+     * Format the waz.task.Task object into String to be saved in the file
+     * @return a formatted waz.task.Task string
      */
     @Override
     public String toDataString() {
