+++ src/main/java/waz/task/Task.java
@@ -1,5 +1,4 @@
-import java.lang.reflect.Array;
-import java.util.ArrayList;
+package waz.task;
 
 public abstract class Task {
     protected String description;
